v2/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
pydantic==2.12.5
python-dotenv==1.2.2
pyjwt==2.11.0
requests==2.32.5
